Refactoring: декомпозировать тему, сервисы
