+++ src/api/services/authService.ts
@@ -1,5 +1,5 @@
-import { IAuthUserRequest, IToken } from "../types";
-import api from "./api";
+import { IAuthUserRequest, IToken } from "../../types";
+import api from "../api";
 
 const authService = {
   login: async (data: IAuthUserRequest): Promise<IToken> => {
