+++ src/app/index.module.scss
@@ -1,4 +1,3 @@
-@import "normalize.css";
 @import "styles/index.scss";
 
 .app {
