+++ src/app/styles/index.scss
@@ -2,3 +2,4 @@
 @import "abstracts/constants.scss";
 @import "abstracts/mixins.scss";
 @import "fonts/index.scss";
+@import "normalize.css";
