+++ src/pages/index.ts
@@ -1,0 +1,8 @@
+import CasePage from "./CasePage/CasePage";
+import Dashboard from "./DashboardPage/Dashboard";
+import DesignersPage from "./DesignersPage/DesignersPage";
+import ErrorPage from "./ErrorPage/ErrorPage";
+import MainPage from "./MainPage/MainPage";
+import ProfilePage from "./ProfilePage/ProfilePage";
+
+export { MainPage, ProfilePage, ErrorPage, DesignersPage, Dashboard, CasePage };
